perf(session): keep dirty flag sticky and fold tests into existing files
Review changes:

The dirty flag is now only cleared after a decode, never after an
encode. Clearing it post-encode could race with a concurrent Set and,
worse, allowed mutations made through previously returned or stored
references (maps, slices, pointers) to be skipped by a later save.
Repeated saves after a write now re-encode, which matches the current
behavior on main.

The unchanged-save branch rechecks the flag under the data lock before
reusing the snapshot, so it synchronizes with mutations the same way the
encode path does.

Tests and benchmarks moved into session_test.go, data_test.go and
middleware_test.go; the three separate files are gone. Added a
regression test that mutates a retained map between two saves. Comments
trimmed per review.
